requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
django-phonenumber-field==7.0.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
phonenumbers==8.12.56
Pillow==9.2.0
psycopg==3.1.3
psycopg2==2.9.3
pycountry==22.3.5
pytz==2022.4
sqlparse==0.4.3
typing-extensions==4.3.0
tzdata==2022.4
